Bugfixes. Now compiles, but database not running properly.

+++ app/src/main/java/com/example/encounter_generator/database/MultipliersEntity.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "encounter_multipliers")
 data class Multipliers (@ColumnInfo(name = "minimum_encounter_size") val minSize: Int,
                         @ColumnInfo(name = "maximum_encounter_size") val maxSize: Int,
-                        @ColumnInfo(name = "encounter_multiplier") val multiplier: Int) {
+                        @ColumnInfo(name = "encounter_multiplier") val multiplier: Double) {
     @PrimaryKey(autoGenerate = true)
     var uid: Int = 0
 }
